[Fix] Don't open a new tab when ctrl+f is pressed
When pressing Ctrl + f to search for something in a form and then clicking e.g. the "search" button, a new tab is opened.

Normally nuCtrlKeyupListener() is called when the Ctrl key is released. This is the case when e.g. Ctrl + c is pressed. But when invoking Ctrl + f a search window is opened (at least in Chrome) and then nuCtrlKeyupListener() is not triggered - probably because the focus is not on the html document anymore.